+++ src/main/java/chess/Controller.java
@@ -1,6 +1,8 @@
 package chess;
 
-import chess.domain.Board;
+import java.util.Scanner;
+import chess.domain.board.Board;
+import chess.domain.board.Coordinate;
 import chess.view.InputView;
 import chess.view.OutputView;
 
@@ -23,6 +25,29 @@ class Controller {
     }
 
     private void play(Board board) {
-        outputView.printBoard(board);
+        while (true) {
+            outputView.printBoard(board);
+            Scanner scanner = new Scanner(System.in);
+
+            String s = scanner.nextLine();
+            String[] split = s.split(" ");
+
+            String source = split[0];
+            String target = split[1];
+
+            String[] sourceInput = source.split(",");
+            int moveStartRank = Integer.parseInt(sourceInput[0]);
+            char moveStartFile = sourceInput[1].toCharArray()[0];
+
+            String[] targetInput = target.split(",");
+            int moveTargetRank = Integer.parseInt(targetInput[0]);
+            char moveTargetFile = targetInput[1].toCharArray()[0];
+
+            try {
+                board.move(new Coordinate(moveStartRank, (char) moveStartFile), new Coordinate(moveTargetRank, (char) moveTargetFile));
+            } catch (Exception e) {
+                System.out.println("[ERROR] : " + e.getMessage());
+            }
+        }
     }
 }
