+++ src/main/java/chess/domain/board/Coordinate.java
@@ -1,4 +1,4 @@
-package chess.domain;
+package chess.domain.board;
 
 import java.util.Objects;
