+++ src/test/java/chess/domain/board/CoordinateTest.java
@@ -1,4 +1,4 @@
-package chess.domain;
+package chess.domain.board;
 
 import static org.assertj.core.api.Assertions.assertThatCode;
