+++ src/test/java/chess/domain/board/RankTest.java
@@ -1,4 +1,4 @@
-package chess.domain;
+package chess.domain.board;
 
 import static org.assertj.core.api.Assertions.assertThatCode;
 import static org.assertj.core.api.Assertions.assertThatThrownBy;
@@ -8,20 +8,20 @@ import org.junit.jupiter.api.Test;
 import org.junit.jupiter.params.ParameterizedTest;
 import org.junit.jupiter.params.provider.ValueSource;
 
-class FileTest {
+class RankTest {
 
     @DisplayName("생성 테스트")
     @Test
     void create() {
-        assertThatCode(() -> new File('a'))
+        assertThatCode(() -> new Rank(1))
                 .doesNotThrowAnyException();
     }
 
-    @DisplayName("열은 소문자 a부터 h까지의 값만 허용한다.")
+    @DisplayName("행은 1부터 8사이의 값만 허용한다.")
     @ParameterizedTest
-    @ValueSource(chars = {'A', 'i'})
-    void checkValue(char source) {
-        assertThatThrownBy(() -> new File(source))
+    @ValueSource(ints = {0, 9})
+    void checkValue(int source) {
+        assertThatThrownBy(() -> new Rank(source))
                 .isInstanceOf(IllegalArgumentException.class);
     }
 }
